Extract stat allocation from stats overview
