+++ website/client/components/userMenu/profileStats.vue
@@ -117,50 +117,19 @@
               li
                 strong {{$t('buffs')}}:
                 | {{user.stats.buffs[stat]}}
-    #allocation(v-if='showAllocation')
-      .row.title-row
-        .col-12.col-md-6
-          h3(v-if='userLevel100Plus', v-once, v-html="$t('noMoreAllocate')")
-          h3
-            | {{$t('statPoints')}}
-            .counter.badge(v-if='user.stats.points || userLevel100Plus')
-              | {{pointsRemaining}}&nbsp;
-        .col-12.col-md-6
-          .float-right
-            toggle-switch(
-              :label="$t('autoAllocation')",
-              v-model='user.preferences.automaticAllocation',
-              @change='setAutoAllocate()'
-            )
-      .row
-        .col-12.col-md-3(v-for='(statInfo, stat) in allocateStatsList')
-          .box.white.row.col-12
-            .col-9
-              div(:class='stat') {{ $t(stats[stat].title) }}
-              .number {{totalAllocatedStats(stat)}}
-              .points {{$t('pts')}}
-            .col-3
-              div
-                .up(v-if='showStatsSave', @click='allocate(stat)')
-              div
-                .down(v-if='showStatsSave', @click='deallocate(stat)')
-      .row.save-row(v-if='showStatsSave')
-        .col-12.col-md-6.offset-md-3.text-center
-          button.btn.btn-primary(@click='saveAttributes()', :disabled='loading') {{ this.loading ?  $t('loading') : $t('save') }}
+    stat-allocation(
+      :user='user',
+      :showAllocation='showAllocation',)
 </template>
 
 <script>
-  import toggleSwitch from 'client/components/ui/toggleSwitch';
   import attributesGrid from 'client/components/inventory/equipment/attributesGrid';
+  import statAllocation from 'client/components/userMenu/statAllocation';
 
   import { mapState } from 'client/libs/store';
   import Content from '../../../common/script/content';
   import { beastMasterProgress, mountMasterProgress } from '../../../common/script/count';
-  import autoAllocate from '../../../common/script/fns/autoAllocate';
-  import allocateBulk from  '../../../common/script/ops/stats/allocateBulk';
   import statsComputed from  '../../../common/script/libs/statsComputed';
-
-  import axios from 'axios';
 
   import size from 'lodash/size';
   import keys from 'lodash/keys';
@@ -170,12 +139,11 @@
   export default {
     props: ['user', 'showAllocation'],
     components: {
-      toggleSwitch,
       attributesGrid,
+      statAllocation,
     },
     data () {
       return {
-        loading: false,
         equipTypes: {
           eyewear: this.$t('eyewear'),
           head: this.$t('headgearCapitalized'),
@@ -186,13 +154,6 @@
           weapon: this.$t('mainHand'),
           _skip: 'skip',
           shield: this.$t('offHand'),
-        },
-
-        allocateStatsList: {
-          str: { title: 'allocateStr', popover: 'strengthText', allocatepop: 'allocateStrPop' },
-          int: { title: 'allocateInt', popover: 'intText', allocatepop: 'allocateIntPop' },
-          con: { title: 'allocateCon', popover: 'conText', allocatepop: 'allocateConPop' },
-          per: { title: 'allocatePer', popover: 'perText', allocatepop: 'allocatePerPop' },
         },
 
         stats: {
@@ -213,12 +174,6 @@
             popover: 'perText',
           },
         },
-        statUpdates: {
-          str: 0,
-          int: 0,
-          con: 0,
-          per: 0,
-        },
         content: Content,
       };
     },
@@ -235,30 +190,16 @@
       statsComputed () {
         return statsComputed(this.user);
       },
-      userLevel100Plus () {
-        return this.user.stats.lvl >= 100;
-      },
-      showStatsSave () {
-        return Boolean(this.user.stats.points);
-      },
-      pointsRemaining () {
-        let points = this.user.stats.points;
-        Object.values(this.statUpdates).forEach(value => {
-          points -= value;
-        });
-        return points;
-      },
-
     },
     methods: {
       getGearTitle (key) {
         return this.flatGear[key].text();
       },
       totalAllocatedStats (stat) {
-        return this.user.stats[stat] + this.statUpdates[stat];
+        return this.user.stats[stat]; // + this.statUpdates[stat];
       },
       totalStatPoints (stat) {
-        return this.statsComputed[stat] + this.statUpdates[stat];
+        return this.statsComputed[stat]; // + this.statUpdates[stat];
       },
       totalCount (objectToCount) {
         let total = size(objectToCount);
@@ -304,56 +245,11 @@
         let display = `${stat}/${totalStat}`;
         return display;
       },
-      allocate (stat) {
-        if (this.pointsRemaining === 0) return;
-        this.statUpdates[stat]++;
-      },
-      deallocate (stat) {
-        if (this.statUpdates[stat] === 0) return;
-        this.statUpdates[stat]--;
-      },
-      async saveAttributes () {
-        this.loading = true;
-
-        const statUpdates = {};
-        ['str', 'int', 'per', 'con'].forEach(stat => {
-          if (this.statUpdates[stat] > 0) statUpdates[stat] = this.statUpdates[stat];
-        });
-
-        // reset statUpdates to zero before request to avoid display errors while waiting for server
-        this.statUpdates = {
-          str: 0,
-          int: 0,
-          con: 0,
-          per: 0,
-        };
-
-        allocateBulk(this.user, { body: { stats: statUpdates } });
-
-        await axios.post('/api/v4/user/allocate-bulk', {
-          stats: statUpdates,
-        });
-
-        this.loading = false;
-      },
-      allocateNow () {
-        autoAllocate(this.user);
-      },
-      setAutoAllocate () {
-        let settings = {
-          'preferences.automaticAllocation': Boolean(this.user.preferences.automaticAllocation),
-          'preferences.allocationMode': 'taskbased',
-        };
-
-        this.$store.dispatch('user:set', settings);
-      },
     },
   };
 </script>
 
 <style lang="scss" scoped>
-  @import '~client/assets/scss/colors.scss';
-
   #stats {
     .box div {
       margin: 0 auto;
@@ -394,74 +290,6 @@
 
   .per {
     color: #4f2a93;
-  }
-
-  #allocation {
-    .title-row {
-      margin-top: 1em;
-      margin-bottom: 1em;
-    }
-
-    .counter.badge {
-      position: relative;
-      top: -0.25em;
-      left: 0.5em;
-      color: #fff;
-      background-color: #ff944c;
-      box-shadow: 0 1px 1px 0 rgba(26, 24, 29, 0.12);
-      width: 24px;
-      height: 24px;
-      border-radius: 50%;
-    }
-
-    .box {
-      width: 148px;
-      height: 84px;
-      padding: .5em;
-      margin: 0 auto;
-
-      div {
-        margin-top: 0;
-      }
-
-      .number {
-        font-size: 40px;
-        text-align: left;
-        color: #686274;
-        display: inline-block;
-      }
-
-      .points {
-        display: inline-block;
-        font-weight: bold;
-        line-height: 1.67;
-        text-align: left;
-        color: #878190;
-        margin-left: .5em;
-      }
-
-      .up, .down {
-        border: solid #a5a1ac;
-        border-width: 0 3px 3px 0;
-        display: inline-block;
-        padding: 3px;
-      }
-
-      .up:hover, .down:hover {
-        cursor: pointer;
-      }
-
-      .up {
-        transform: rotate(-135deg);
-        -webkit-transform: rotate(-135deg);
-        margin-top: 1em;
-      }
-
-      .down {
-        transform: rotate(45deg);
-        -webkit-transform: rotate(45deg);
-      }
-    }
   }
 
   #attributes {
@@ -518,9 +346,5 @@
 
   .mount {
     margin-top: -0.2em !important;
-  }
-
-  .save-row {
-    margin-top: 1em;
   }
 </style>
